requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.11
email_validator==2.0.0
google-auth==2.22.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.0
google-api-python-client==2.104.0
